move SplashActivity from utils to it's own package, main
