+++ app/src/main/java/org/kiwix/kiwixmobile/main/SplashActivity.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.utils;
+package org.kiwix.kiwixmobile.main;
 
 import android.content.Context;
 import android.content.Intent;
